finished adding debit card to front end and cleaning up code
